App users, JWT Authentication, permissions, API manual tests, visibility and company centralized signup developed
The users model was separated from companies and put into a new users app where is executed all the users logic and the main authentication parts. Also there was implemented the JWT Authentication with the login and the refresh endpoint, and the permissions with the new auth system was changed all over the app. The visibility was implemented for manage soft deleting in the critical data all over the application. Finaly the supplier and buyer signups were deleted to unify in a unique company signup with supplier and profile activations